[feat: gw api] Add common listener config for gateway api

+++ pkg/gateway/routeutils/loader.go
@@ -47,7 +47,7 @@ var L7RouteFilter LoadRouteFilter = &routeFilterImpl{
 // Loader will load all data Kubernetes that are pertinent to a gateway (Routes, Services, Target Group Configurations).
 // It will output the data using a map which maps listener port to the various routing rules for that port.
 type Loader interface {
-	LoadRoutesForGateway(ctx context.Context, gw gwv1.Gateway, filter LoadRouteFilter) (map[int][]RouteDescriptor, error)
+	LoadRoutesForGateway(ctx context.Context, gw gwv1.Gateway, filter LoadRouteFilter) (map[int32][]RouteDescriptor, error)
 }
 
 var _ Loader = &loaderImpl{}
@@ -69,7 +69,7 @@ func NewLoader(k8sClient client.Client, logger logr.Logger) Loader {
 }
 
 // LoadRoutesForGateway loads all relevant data for a single Gateway.
-func (l *loaderImpl) LoadRoutesForGateway(ctx context.Context, gw gwv1.Gateway, filter LoadRouteFilter) (map[int][]RouteDescriptor, error) {
+func (l *loaderImpl) LoadRoutesForGateway(ctx context.Context, gw gwv1.Gateway, filter LoadRouteFilter) (map[int32][]RouteDescriptor, error) {
 	// 1. Load all relevant routes according to the filter
 	loadedRoutes := make([]preLoadRouteDescriptor, 0)
 	for route, loader := range l.allRouteLoaders {
@@ -97,12 +97,12 @@ func (l *loaderImpl) LoadRoutesForGateway(ctx context.Context, gw gwv1.Gateway, 
 }
 
 // loadChildResources responsible for loading all resources that a route descriptor references.
-func (l *loaderImpl) loadChildResources(ctx context.Context, preloadedRoutes map[int][]preLoadRouteDescriptor) (map[int][]RouteDescriptor, error) {
+func (l *loaderImpl) loadChildResources(ctx context.Context, preloadedRoutes map[int][]preLoadRouteDescriptor) (map[int32][]RouteDescriptor, error) {
 	// Cache to reduce duplicate route look ups.
 	// Kind -> [NamespacedName:Previously Loaded Descriptor]
 	resourceCache := make(map[string]RouteDescriptor)
 
-	loadedRouteData := make(map[int][]RouteDescriptor)
+	loadedRouteData := make(map[int32][]RouteDescriptor)
 
 	for port, preloadedRouteList := range preloadedRoutes {
 		for _, preloadedRoute := range preloadedRouteList {
@@ -112,7 +112,7 @@ func (l *loaderImpl) loadChildResources(ctx context.Context, preloadedRoutes map
 
 			cachedRoute, ok := resourceCache[cacheKey]
 			if ok {
-				loadedRouteData[port] = append(loadedRouteData[port], cachedRoute)
+				loadedRouteData[int32(port)] = append(loadedRouteData[int32(port)], cachedRoute)
 				continue
 			}
 
@@ -120,7 +120,7 @@ func (l *loaderImpl) loadChildResources(ctx context.Context, preloadedRoutes map
 			if err != nil {
 				return nil, err
 			}
-			loadedRouteData[port] = append(loadedRouteData[port], generatedRoute)
+			loadedRouteData[int32(port)] = append(loadedRouteData[int32(port)], generatedRoute)
 			resourceCache[cacheKey] = generatedRoute
 		}
 	}
